+++ naruto-1/build.gradle
@@ -9,8 +9,8 @@ version '1.0-SNAPSHOT'
 repositories {
     mavenCentral()
 }
-sourceSets.main.java.srcDirs = ['src']
-sourceSets.test.java.srcDirs = ['test']
+//sourceSets.main.java.srcDirs = ['src']
+//sourceSets.test.java.srcDirs = ['test']
 
 
 dependencies {
