+++ naruto-1/src/tipos_ninja/NinjaDeNinjutsu.java
@@ -1,8 +1,8 @@
-package naruto.src.tipos_ninja;
+package tipos_ninja;
 
 
-import naruto.src.interfaces.NinjaInterface;
-import naruto.src.personagem.Personagem;
+import interfaces.NinjaInterface;
+import personagem.Personagem;
 
 import java.util.List;
