+++ naruto-1/test/tipos_ninja/NinjaDeNinjutsuTest.java
@@ -1,7 +1,7 @@
-package src.naruto.src.tipos_ninja;
+package tipos_ninja;
 
 import jdk.jfr.Label;
-import naruto.src.tipos_ninja.NinjaDeTaijutsu;
+import tipos_ninja.NinjaDeNinjutsu;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -9,16 +9,16 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class NinjaDeTaijutsuTest {
+public class NinjaDeNinjutsuTest {
 
     @Test
     @Label("Adiciona um novo jutsu ao array de Jutsu")
     public void deveUsarJutsu(){
         List<String> jutsus = new ArrayList<>();
         jutsus.add("Suiton");
-        NinjaDeTaijutsu ninjaDeTaijutsu = new NinjaDeTaijutsu("Mateus",28, "tres",
-                jutsus,5 );
-        assertEquals("Usando o Taijutsu",ninjaDeTaijutsu.usarJutsu());
+        NinjaDeNinjutsu ninjaDeNinjutsu = new NinjaDeNinjutsu("Mateus",28, "tres",
+                                                                jutsus,5 );
+        assertEquals("Usando o Ninjutsu",ninjaDeNinjutsu.usarJutsu());
     }
 
     @Test
@@ -26,7 +26,7 @@ public class NinjaDeTaijutsuTest {
     public void deveDesviar(){
         List<String> jutsus = new ArrayList<>();
         jutsus.add("Fuuton");
-        NinjaDeTaijutsu ninjaDeTaijutsu = new NinjaDeTaijutsu("Mateus",28, "tres", jutsus,15);
-        assertEquals("Desviou para: Lado", ninjaDeTaijutsu.desviar("Lado"));
+        NinjaDeNinjutsu ninjaDeNinjutsu = new NinjaDeNinjutsu("Mateus",28, "tres", jutsus,9);
+        assertEquals("Desviou para: Frente", ninjaDeNinjutsu.desviar("Frente"));
     }
 }
